+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # byteranges
 
+## 1.1.0 (2023-02-12)
+
+- Fix return type of parse function (parts cannot be unsatisfied ranges).
+- Add isSatisfiableRange, isCompleteLengthRange, and isCompleteLengthRangeParts type guard functions.
+
 ## 1.0.1 (2023-01-21)
 
 - Fix package.json.